+++ CHANGES.md
@@ -12,6 +12,9 @@ For change history for [MOST][3], see [most/CHANGES.md](most/CHANGES.md).
 Since last release
 ------------------
 
+#### 4/19/18
+  - Add AC OPF tests for cases with ref bus ~= 1 and ref angle ~= 0.
+
 #### 4/12/18
   - Add to `nested_struct_copy()` ability to copy fields that are struct
     arrays.